cards_binders/requirements.txt
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.24.0
python-dateutil>=2.8.0
urllib3>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0
